refactor structure of project and fix the errors
