+++ src/main/kotlin/domain/usecase/project/CreateProjectUseCase.kt
@@ -1,8 +1,8 @@
-package com.berlin.logic.usecase.project
+package com.berlin.domain.usecase.project
 
-import com.berlin.logic.generateIdHelper.IdGenerator
-import com.berlin.logic.repositories.ProjectRepository
-import com.berlin.model.Project
+import com.berlin.domain.helper.IdGenerator
+import com.berlin.domain.repository.ProjectRepository
+import com.berlin.domain.model.Project
 
 class CreateProjectUseCase(
     private val projectRepository: ProjectRepository,
