+++ src/main/kotlin/domain/usecase/project/DeleteProjectUseCase.kt
@@ -1,6 +1,6 @@
-package com.berlin.logic.usecase.project
+package com.berlin.domain.usecase.project
 
-import com.berlin.logic.repositories.ProjectRepository
+import com.berlin.domain.repository.ProjectRepository
 
 class DeleteProjectUseCase (
     private val projectRepository: ProjectRepository
