--- src/main/kotlin/logic/generateIdHelper/IdGenerator.kt
@@ -1,11 +1,0 @@
-package com.berlin.logic.generateIdHelper
-
-import kotlin.random.Random
-
-interface IdGenerator {
-    fun generateId(
-        prefix: String,
-        padChar: Char = prefix[prefix.length / 2],
-        padCharLength: Int  = prefix.length + Random.nextInt(1, 5)
-    ): String
-}
