--- src/main/kotlin/logic/repositories/TaskRepository.kt
@@ -1,13 +1,0 @@
-package com.berlin.logic.repositories
-
-import com.berlin.model.Task
-import com.berlin.model.User
-
-interface TaskRepository {
-    fun createTask(task:Task):Boolean
-    fun getTaskById(taskId:String):Task?
-    fun getTasksByProjectId(projectId:String):List<Task>
-    fun updateTask(task: Task):Boolean
-    fun deleteTaskById(taskId: String):Boolean
-    fun getAssignedUserByTaskId(taskId: String):User?
-}
