--- src/main/kotlin/model/Project.kt
@@ -1,9 +1,0 @@
-package com.berlin.model
-
-data class Project(
-    val id:String,
-    val name:String,
-    val description:String?,
-    val statesId:List<String>?,
-    val tasksId:List<String>?
-)
