--- src/main/kotlin/model/User.kt
@@ -1,10 +1,0 @@
-package com.berlin.model
-
-import com.berlin.domain.model.UserRole
-
-data class User(
-    val id:String,
-    val userName:String,
-    val password:String,
-    val role: UserRole
-)
