+++ src/test/kotlin/logic/usecase/project/GetProjectByIdUseCaseTest.kt
@@ -1,15 +1,14 @@
 package logic.usecase.project;
 
 import com.berlin.helper.projectHelper
-import com.berlin.logic.repositories.ProjectRepository
-import com.berlin.logic.usecase.project.GetProjectByIdUseCase
+import com.berlin.domain.repository.ProjectRepository
+import com.berlin.domain.usecase.project.GetProjectByIdUseCase
 import com.google.common.truth.Truth.assertThat
 import io.mockk.every
 import io.mockk.mockk
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.CsvSource
 import org.junit.jupiter.params.provider.ValueSource
 import kotlin.test.Test
